Implement full test

+++ analisi.sh
@@ -18,10 +18,11 @@ CLIENT_PRODS=$(echo "$CLIENT_LINES" | grep "prodotti acquistati" | sed 's/.*=\s/
 CLIENT_TOT_TS=$(echo "$CLIENT_LINES" | grep "tempo totale" | sed 's/.*=\s//')
 CLIENT_QUEUE_TS=$(echo "$CLIENT_LINES" | grep "tempo in coda" | sed 's/.*=\s//')
 CLIENT_CHANGES=$(echo "$CLIENT_LINES" | grep "cambi di coda" | sed 's/.*=\s//')
+CLIENTS=$(echo "$CLIENT_IDS" | tail -1)
 
 # Get number of clients
-echo -n "Numero di clienti: "
-echo "$CLIENT_IDS" | tail -1
+# echo -n "Numero di clienti: "
+# echo $CLIENTS
 
 echo "$CLIENT_IDS" > .ids
 echo "$CLIENT_PRODS" > .prods
@@ -29,10 +30,35 @@ echo "$CLIENT_TOT_TS" > .tots
 echo "$CLIENT_QUEUE_TS" > .qts
 echo "$CLIENT_CHANGES" > .changes
 
-echo RIASSUNTO CLIENTI
-echo -e "ID\tPROD.\tT.TOT\tT.CODA\tN.CODE"
-paste .ids .prods .tots .qts .changes
+# echo RIASSUNTO CLIENTI
+# echo -e "ID\tPROD.\tT.TOT\tT.CODA\tN.CODE"
+paste .ids .prods .tots .qts .changes > analisi.log
 
+# Imposta il locale in modo da intepretare i numeri in virgola mobile
+# correttamente 
+export LC_ALL=C
+
+# # Fallisce se trova numeri negativi
+[ $(grep -q -s -- '-[1-9]*' analisi.log) ] &&
+  { echo "Sono presenti valori negativi" >> /dev/stderr; exit 1; }
+
+# Fallisce se un cliente acquista troppi prodotti
+[ -z "$(awk '$1 > 100' .prods)" ] || 
+  { echo "errore prodotti acquistati dai clienti" >> /dev/stderr; exit 1; }
+
+# Fallisce se un cliente impiega troppo tempo nel supermercato
+[ -z "$(awk '$1 > 20' .tots)" ] || 
+  { echo "errore tempo totale cliente" >> /dev/stderr; exit 1; }
+
+# Fallisce se un cliente impiega troppo tempo in coda
+[ -z "$(awk '$1 > 20' .qts)" ] || 
+  { echo "errore tempo in coda cliente" >> /dev/stderr; exit 1; }
+
+# Fallisce se un cliente cambia coda troppe volte
+[ -z "$(awk '$1 > 6' .changes)" ] || 
+  { echo "errore numero cambio code" >> /dev/stderr; exit 1; }
+
+# Rimuove i file temporanei
 rm -f .ids .prods .tots .qts .changes
 
 # Cashier lines sorted by ID in ascending order
@@ -51,17 +77,26 @@ echo "$CASH_TOT_TS" > .tots
 echo "$CASH_AVG_TS" > .avgs
 echo "$CASH_CLOSE" > .close
 
-echo
-echo RIASSUNTO CASSIERI
-echo -e "ID\tPROD.\tCLIENTI\tT.TOT\tT.MEDIO\tN.CHIUS."
-paste .ids .prods .clients .tots .avgs .close
+# echo
+# echo RIASSUNTO CASSIERI
+# echo -e "ID\tPROD.\tCLIENTI\tT.TOT\tT.MEDIO\tN.CHIUS."
+paste .ids .prods .clients .tots .avgs .close > analisi.log
+
+# # Fallisce se trova numeri negativi
+[ $(grep -q -s -- '-[1-9]*' analisi.log) ] &&
+  { echo "Sono presenti valori negativi" >> /dev/stderr; exit 1; }
+
+# Fallisce se sono stati serviti troppi clienti
+[ -z "$(awk '$1 > $CLIENTS' .clients)" ] || { echo "errore clienti serviti" >> /dev/stderr; exit 1; }
+
+# Fallisce se le casse sono state aperte troppo tempo
+[ -z "$(awk '$1 > 50' .tots)" ] || { echo "errore tempo totale cassiere" >> /dev/stderr; exit 1; }
+
+# Fallisce se le casse hanno un tempo medio di servizio troppo alto
+[ -z "$(awk '$1 > 2' .avgs)" ] || { echo "errore tempo medio di servizio" >> /dev/stderr; exit 1; }
+
+# Fallisce se le casse sono state chiuse troppe volte
+[ -z "$(awk '$1 > 3' .close)" ] || { echo "errore numero di chiusure casse" >> /dev/stderr; exit 1; }
 
 rm -f .ids .prods .clients .tots .avgs .close
-
-
-
-
-
-
-
